Added get/setPlayheadTime(); iOS check for cloud items
New functions can read and set the playhead (playback position) of the
currently playing song. The playhead value is a number in seconds.

Also pickSong() now takes a param 'downloadedSongsOnly'. The default is
true, and causes the iOS device to only show songs that have been
downloaded to the device. (Not implemented on Android).